Cloud Computing/connect-appengine-cloudsql/requirements.txt
flask==2.1.0
passlib==1.7.4
PyMySQL==1.0.2
psycopg2==2.8.4
#psycopg2-binary==2.9.3 # you will need either the binary or the regular - for more info see http://initd.org/psycopg/docs/install.html
SQLAlchemy==1.4.37
gunicorn==20.1.0
Flask-RESTful==0.3.8
PyJWT==1.7.1
Flask-SQLAlchemy==2.4.1